V1

+++ src/main/java/com/tpe/domain/Role.java
@@ -14,19 +14,21 @@ import javax.persistence.*;
 @NoArgsConstructor
 
 @Entity
-@Table(name = "tbl_role")
+@Table(name="tbl_role")
 public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Enumerated(EnumType.STRING)
-    @Column(length = 30,nullable = false)
+    @Column(length = 30, nullable = false)
     private UserRole name;
 
     @Override
     public String toString() {
-        return "Role{"+"name="+name+'}';
+        return "Role{" +
+                "name=" + name +
+                '}';
     }
 }
